+++ packages/tooling/fast-tooling/src/web-components/css-layout/css-layout.template.ts
@@ -1,4 +1,5 @@
 import { html, repeat } from "@microsoft/fast-element";
+import { ElementDefinitionContext, Switch } from "@microsoft/fast-foundation";
 import { CSSLayout } from "./css-layout";
 
 /**
@@ -533,12 +534,15 @@ const columnGap = html`
  * The template for the css-layout component.
  * @public
  */
-export const CSSLayoutTemplate = html<CSSLayout>`
+export const CSSLayoutTemplate = (context: ElementDefinitionContext) => html<CSSLayout>`
     <template>
         <div class="control-region">
-            <fast-switch @change="${(x, c) => x.handleToggleCSSLayout()}">
+            <${context.tagFor(Switch)} checked="${x => x.flexEnabled}" @change="${(
+    x,
+    c
+) => x.handleToggleCSSLayout()}">
                 Enable Flexbox
-            </fast-switch>
+            </${context.tagFor(Switch)}>
         </div>
         <div class="flexbox-controls${x => (x.flexEnabled ? ` active` : "")}">
             <div class="control-region">
@@ -565,7 +569,7 @@ export const CSSLayoutTemplate = html<CSSLayout>`
                                     aria-label="${x => x}"
                                     name="${(x, c) => c.parent.flexDirectionName}"
                                     value="${x => x}"
-                                    ?selected="${(x, c) =>
+                                    ?checked="${(x, c) =>
                                         c.parent.flexDirectionValue === x}"
                                     @change="${(x, c) =>
                                         c.parent.handleCSSChange(
@@ -606,7 +610,7 @@ export const CSSLayoutTemplate = html<CSSLayout>`
                                     aria-label="${x => x}"
                                     name="${(x, c) => c.parent.justifyContentName}"
                                     value="${x => x}"
-                                    ?selected="${(x, c) =>
+                                    ?checked="${(x, c) =>
                                         c.parent.justifyContentValue === x}"
                                     @change="${(x, c) =>
                                         c.parent.handleCSSChange(
@@ -649,7 +653,7 @@ export const CSSLayoutTemplate = html<CSSLayout>`
                                     aria-label="${x => x}"
                                     name="${(x, c) => c.parent.alignContentName}"
                                     value="${x => x}"
-                                    ?selected="${(x, c) =>
+                                    ?checked="${(x, c) =>
                                         c.parent.alignContentValue === x}"
                                     @change="${(x, c) =>
                                         c.parent.handleCSSChange(
@@ -687,8 +691,7 @@ export const CSSLayoutTemplate = html<CSSLayout>`
                                     aria-label="${x => x}"
                                     name="${(x, c) => c.parent.alignItemsName}"
                                     value="${x => x}"
-                                    ?selected="${(x, c) =>
-                                        c.parent.alignItemsValue === x}"
+                                    ?checked="${(x, c) => c.parent.alignItemsValue === x}"
                                     @change="${(x, c) =>
                                         c.parent.handleCSSChange(
                                             "alignItemsValue",
@@ -709,9 +712,11 @@ export const CSSLayoutTemplate = html<CSSLayout>`
                                 ${rowGap}
                             </div>
                             <input
+                                name="${x => x.rowGapName}"
                                 class="css-row-gap"
                                 type="number"
                                 id="fast-tooling-css-row-gap"
+                                value="${x => x.rowGapValue}"
                                 @input="${(x, c) =>
                                     x.handleCSSChange("rowGapValue", c.event)}"
                             />
@@ -724,9 +729,11 @@ export const CSSLayoutTemplate = html<CSSLayout>`
                                 ${columnGap}
                             </div>
                             <input
+                                name="${x => x.columnGapName}"
                                 class="css-column-gap"
                                 type="number"
                                 id="fast-tooling-css-column-gap"
+                                value="${x => x.columnGapValue}"
                                 @input="${(x, c) =>
                                     x.handleCSSChange("columnGapValue", c.event)}"
                             />
@@ -757,7 +764,7 @@ export const CSSLayoutTemplate = html<CSSLayout>`
                                     aria-label="${x => x}"
                                     name="${(x, c) => c.parent.flexWrapName}"
                                     value="${x => x}"
-                                    ?selected="${(x, c) => c.parent.flexWrapValue === x}"
+                                    ?checked="${(x, c) => c.parent.flexWrapValue === x}"
                                     @change="${(x, c) =>
                                         c.parent.handleCSSChange(
                                             "flexWrapValue",