+++ packages/web-components/fast-foundation/src/checkbox/checkbox.form-associated.ts
@@ -1,7 +1,7 @@
-import { FASTElement } from "@microsoft/fast-element";
 import { FormAssociated } from "../form-associated/form-associated";
+import { FoundationElement } from "../foundation-element";
 
-class _Checkbox extends FASTElement {}
+class _Checkbox extends FoundationElement {}
 interface _Checkbox extends FormAssociated {}
 
 /**