+++ packages/web-components/fast-foundation/src/switch/switch.template.ts
@@ -1,12 +1,19 @@
 import { html, slotted } from "@microsoft/fast-element";
 import type { ViewTemplate } from "@microsoft/fast-element";
-import type { Switch } from "./switch";
+import type { Switch, SwitchOptions } from "./switch";
+import type { ElementDefinitionContext } from "../design-system";
 
 /**
  * The template for the {@link @microsoft/fast-foundation#(Switch:class)} component.
  * @public
  */
-export const SwitchTemplate: ViewTemplate<Switch> = html`
+export const switchTemplate: (
+    context: ElementDefinitionContext,
+    definition: SwitchOptions
+) => ViewTemplate<Switch> = (
+    context: ElementDefinitionContext,
+    definition: SwitchOptions
+) => html`
     <template
         role="switch"
         aria-checked="${x => x.checked}"
@@ -27,7 +34,7 @@ export const SwitchTemplate: ViewTemplate<Switch> = html`
             <slot ${slotted("defaultSlottedNodes")}></slot>
         </label>
         <div part="switch" class="switch">
-            <span class="checked-indicator" part="checked-indicator"></span>
+            <slot name="switch">${definition.switch || ""}</slot>
         </div>
         <span class="status-message" part="status-message">
             <span class="checked-message" part="checked-message">