+++ packages/web-components/fast-foundation/src/tab/tab.template.ts
@@ -1,12 +1,20 @@
 import { html } from "@microsoft/fast-element";
 import type { ViewTemplate } from "@microsoft/fast-element";
 import type { Tab } from "./tab";
+import type { FoundationElementDefinition } from "../foundation-element";
+import type { ElementDefinitionContext } from "../design-system";
 
 /**
  * The template for the {@link @microsoft/fast-foundation#Tab} component.
  * @public
  */
-export const TabTemplate: ViewTemplate<Tab> = html`
+export const tabTemplate: (
+    context: ElementDefinitionContext,
+    definition: FoundationElementDefinition
+) => ViewTemplate<Tab> = (
+    context: ElementDefinitionContext,
+    definition: FoundationElementDefinition
+) => html`
     <template slot="tab" role="tab" aria-disabled="${x => x.disabled}">
         <slot></slot>
     </template>